Styling update
