+++ html/css/style.css
@@ -1,3 +1,56 @@
+:root {
+    /* Typography */
+    --font-primary: "Exo 2", sans-serif;
+    --font-weight-regular: 400;
+    --font-weight-medium: 500;
+    --font-weight-bold: 700;
+    --font-weight-light: 300;
+
+    /* Colors */
+    --md-primary: #0061a4;
+    --md-on-primary: #ffffff;
+    --md-primary-container: #d1e4ff;
+    --md-on-primary-container: #001d36;
+
+    --md-error: #ba1a1a;
+    --md-on-error: #ffffff;
+    --md-error-container: #ffdad6;
+    --md-on-error-container: #410002;
+
+    --md-success: #006e1c;
+    --md-on-success: #ffffff;
+    --md-success-container: #9df996;
+    --md-on-success-container: #002106;
+
+    --md-warning: #7d5800;
+    --md-on-warning: #ffffff;
+    --md-warning-container: #ffdf93;
+    --md-on-warning-container: #271900;
+
+    --md-info: #0062a1;
+    --md-on-info: #ffffff;
+    --md-info-container: #d0e4ff;
+    --md-on-info-container: #001d35;
+
+    /* Surface colors */
+    --md-surface: #fcfcff;
+    --md-on-surface: #1a1c1e;
+    --md-surface-variant: #dfe2eb;
+    --md-on-surface-variant: #43474e;
+
+    /* Elevation */
+    --md-elevation-1: 0px 1px 3px 1px rgba(0, 0, 0, 0.15);
+
+    /* Border Radius */
+    --md-radius-small: 4px;
+    --md-radius-medium: 8px;
+}
+
+* {
+    font-family: var(--font-primary);
+    font-weight: var(--font-weight-regular);
+}
+
 html::-webkit-scrollbar {
     display: none;
 }
@@ -5,77 +58,91 @@ html::-webkit-scrollbar {
 @media (width: 3840px) and (height: 2160px) {
     .success {
         background-color: rgba(23, 23, 23, 90%);
-        border-radius: 10px;
-        box-shadow: 0rem 0rem 0.1rem 0.05rem #000000;
-        border-left: 0.5rem solid #20bb44;
+        color: var(--md-on-surface);
+        box-shadow: var(--md-elevation-1);
+        border-left: 0.5rem solid var(--md-success);
         font-size: 1.5vh;
     }
 
     .primary {
         background-color: rgba(23, 23, 23, 90%);
-        border-radius: 10px;
-        box-shadow: 0rem 0rem 0.1rem 0.05rem #000000;
-        border-left: 5px solid #1c75d2;
+        color: var(--md-on-surface);
+        box-shadow: var(--md-elevation-1);
+        border-left: 5px solid var(--md-primary);
         font-size: 1.5vh;
     }
 
     .error {
         background-color: rgba(23, 23, 23, 90%);
-        border-radius: 10px;
-        box-shadow: 0rem 0rem 0.1rem 0.05rem #000000;
-        border-left: 5px solid #c10114;
+        color: var(--md-on-surface);
+        box-shadow: var(--md-elevation-1);
+        border-left: 5px solid var(--md-error);
+        font-size: 1.5vh;
+    }
+
+    .warning {
+        background-color: rgba(23, 23, 23, 90%);
+        color: var(--md-on-surface);
+        box-shadow: var(--md-elevation-1);
+        border-left: 5px solid var(--md-warning);
         font-size: 1.5vh;
     }
 
     .police {
         background-color: rgba(23, 23, 23, 90%);
-        border-radius: 10px;
-        box-shadow: 0rem 0rem 0.1rem 0.05rem #000000;
-        border-left: 5px solid #2197f2;
+        color: var(--md-on-surface);
+        box-shadow: var(--md-elevation-1);
+        border-left: 5px solid var(--md-info);
         font-size: 1.5vh;
     }
 
     .ambulance {
         background-color: rgba(23, 23, 23, 90%);
-        border-radius: 10px;
-        box-shadow: 0rem 0rem 0.1rem 0.05rem #000000;
-        border-left: 5px solid #f44236;
+        color: var(--md-on-surface);
+        box-shadow: var(--md-elevation-1);
+        border-left: 5px solid var(--md-error);
         font-size: 1.5vh;
     }
 }
 
 .success {
-    background-color: #20bb44;
-    border-radius: 5px;
-    border: 1px solid #6ae587;
+    background-color: var(--md-success-container);
+    color: var(--md-on-success-container);
+    padding: 12px 16px;
+    font-weight: var(--font-weight-medium);
 }
 
 .primary {
-    background-color: #1c75d2;
-    border-radius: 5px;
-    border: 1px solid #6facec;
+    background-color: var(--md-primary-container);
+    color: var(--md-on-primary-container);
+    padding: 12px 16px;
+    font-weight: var(--font-weight-medium);
 }
 
 .warning {
-    background-color: #ff9800;
-    border-radius: 5px;
-    border: 1px solid #ffc107;
+    background-color: var(--md-warning-container);
+    color: var(--md-on-warning-container);
+    padding: 12px 16px;
+    font-weight: var(--font-weight-medium);
 }
 
 .error {
-    background-color: #c10114;
-    border-radius: 5px;
-    border: 1px solid #fe4255;
+    background-color: var(--md-error-container);
+    color: var(--md-on-error-container);
+    padding: 12px 16px;
+    font-weight: var(--font-weight-medium);
 }
 
 .police {
-    background-color: #2197f2;
-    border-radius: 5px;
-    border: 1px solid #7ac1f7;
+    background-color: var(--md-info-container);
+    color: var(--md-on-info-container);
+    padding: 12px 16px;
+    font-weight: var(--font-weight-medium);
 }
 
 .ambulance {
-    background-color: #f44236;
-    border-radius: 5px;
-    border: 1px solid #f88e86;
+    background-color: var(--md-error-container);
+    color: var(--md-on-error-container);
+    padding: 12px 16px;
+    font-weight: var(--font-weight-medium);
 }
