+++ html/index.html
@@ -1,6 +1,6 @@
 <html>
     <head>
-        <link href="https://fonts.googleapis.com/css2?family=Poppins&display=swap" rel="stylesheet" />
+        <link href="https://fonts.googleapis.com/css2?family=Exo+2:wght@300;400;500;600;700&display=swap" rel="stylesheet" />
         <link href="https://cdn.jsdelivr.net/npm/quasar@2.12.0/dist/quasar.prod.css" rel="stylesheet" type="text/css" />
         <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css" />
         <link href="https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700,900|Material+Icons" rel="stylesheet" type="text/css" />
@@ -11,7 +11,7 @@
         <script src="js/drawtext.js"></script>
         <link rel="stylesheet" href="css/drawtext.css" />
     </head>
-    <body style="font-family: 'Poppins', sans-serif">
+    <body>
         <div id="q-app" style="min-height: 100vh"></div>
         <div id="drawtext-container">
             <div id="text" class="text"></div>
