+++ html/js/config.js
@@ -1,45 +1,53 @@
-import { fetchNui, isEnvBrowser } from "./utils.js";
-import { BrowserMockConfigData } from "./testing.js";
-
-export const DEV_MODE = false;
-
-/**
- * @typedef NotiVariantData
- * @property {string} icon
- * @property {string} color
- **/
-
-/**
- * Will hold config statically outside of Vue state
- * @property {Record<string, NotiVariantData>} VariantDefinitions
- * @property {Record<string, any>} NotificationStyling
- **/
 export let NOTIFY_CONFIG = null;
 
-/**
- * Pure function taking a notification type and returning an object
- * with style details
- * @param {string} variant
- * @returns NotiVariantData
- **/
+const defaultConfig = {
+    NotificationStyling: {
+        group: true,
+        position: "top-right",
+        progress: true,
+    },
+    VariantDefinitions: {
+        success: {
+            classes: "success",
+            icon: "done",
+        },
+        primary: {
+            classes: "primary",
+            icon: "info",
+        },
+        error: {
+            classes: "error",
+            icon: "dangerous",
+        },
+        police: {
+            classes: "police",
+            icon: "local_police",
+        },
+        ambulance: {
+            classes: "ambulance",
+            icon: "fas fa-ambulance",
+        },
+    },
+};
+
 export const determineStyleFromVariant = (variant) => {
     const variantData = NOTIFY_CONFIG.VariantDefinitions[variant];
     if (!variantData) throw new Error(`Style of type: ${variant}, does not exist in the config`);
     return variantData;
 };
 
-// Fetch and set NOTIFY_CONFIG from client script callback
 export const fetchNotifyConfig = async () => {
-    NOTIFY_CONFIG = await fetchNui("getNotifyConfig", {}, BrowserMockConfigData);
-    if (isEnvBrowser() || DEV_MODE) {
-        console.log("Fetched Config:");
-        console.dir(NOTIFY_CONFIG);
+    try {
+        NOTIFY_CONFIG = await window.fetchNui("getNotifyConfig", {});
+        if (!NOTIFY_CONFIG) {
+            NOTIFY_CONFIG = defaultConfig;
+        }
+    } catch (error) {
+        console.error("Failed to fetch notification config, using default", error);
+        NOTIFY_CONFIG = defaultConfig;
     }
 };
 
-// We specifically wait for all other files to load
-// just in case of a race condition between client handlers
-// and NUI fetch call
 window.addEventListener("load", async () => {
     await fetchNotifyConfig();
 });
